# Conflicts:
#	backend/Game/src/main/java/com/maeng/game/domain/room/service/RoomService.java
#	backend/Game/src/main/java/com/maeng/game/global/config/RabbitConfig.java

+++ backend/Game/src/main/java/com/maeng/game/domain/jwac/dto/JwacBidInfoDto.java
@@ -9,5 +9,5 @@ public class JwacBidInfoDto {
 	private String gameCode;
 	private String nickname;
 	private int round;
-	private int bidAmount;
+	private long bidAmount;
 }
